Ensure placement request has upcoming booking
The test in question relies on a Placement Request created from a
factory, and checks for the presence of a 'Change placement' action on
the page. This action only shows if the placement status is not
'departed' or 'notArrived'. However, the factory would set the status
randomly, with one of those 2 occurring approximately 18% of the time
(2/11 possible statuses).

This ensures the placement request being tested has an upcoming booking,
rather than one of any status.