Always show action column on OOSB list
The route itself is only accessible to users with the permission to view
OOSB, so the link can always be displayed. The permission to create OOSB
defines whether the actual actions (create/update) are displayed on each
page. This means a user with the view permission can still get details
about an OOSB.