Remove uses of Matcher role
This role should be deprecated. The only actual use, to fetch a list of
users that can be assigned tasks (either assessments or requests for
placement) has been replaced with the Appeals manager role, as it is
more relevant and covers the same users.

All other uses are for tests, where it has also been replaced with
Appeals manager so it can be fully removed in the future.