Improve Record departure test
The test in question was actively avoiding checking the 'Breach or
recall' radio option, to avoid needing more complicated logic for the
rest of the test. This however wasn't avoiding some of the options that
show the 'Move on' page, resulting in some flaky tests.

The flakiness has been fixed, and the test has been improved by allowing
a random radio button to be selected, and acting in the relevant way
with further screens being shown.