Use Application document renderer for assessment
This replaces instances where what should be a read-only version of a
submitted assessment, based on its `document` property, was instead
being rendered by iterating over the submitted `data` to render the
answers again. This was an issue for questions where the answers have
changed, as those had to be handled with special cases in the `data`
iterator approach.

Using this, the document is rendered exactly as it was saved on
submissions, removing the need for those special cases. Adding/removing
questions will no longer affect previously submitted assessments.