Use separate callable workflow for test deployment
This separate workflow is run on a `workflow_dispatch` action, which
means it can be triggered from the GitHub Actions UI for any given
branch. This means we are now able to deploy branches to test in order
to showcase work, or to conduct user research on work in progress
without having to first merge it.