Add leading zero to time input where required
This avoids the API returning an error when the user hasn't entered the
leading zero. We do this instead of enforcing a leading zero on
validation, to avoid adding friction to user input.

The E2E tests now remove the leading zero from date and time entry for a
more realistic user input.