Add description for sortable table functionality
Each table should have a caption clearly stating that column with
buttons can be sorted. This is impractical with the components we have,
but another approach that seems to be rather equivalent [0] is to add
this description in an `aria-describedby` attribute. The element holding
this description has been added to the main body near the footer, as it
shouldn't hurt for it to be present on all pages.

[0] https://www.w3.org/WAI/tutorials/tables/caption-summary/#using-aria-describedby-to-provide-a-table-summary