Add description for sortable table functionality
Each table should have a caption clearly stating that column with
buttons can be sorted. This is impractical with the components we have,
but another approach that seems to be rather equivalent [0] is to add
this description in an `aria-describedby` attribute.

This commit introduces a new macro for sortable tables, which
automatically adds an element describing the sorting funcionality before
the table and adds the relevant attribute to the table.

[0] https://www.w3.org/WAI/tutorials/tables/caption-summary/#using-aria-describedby-to-provide-a-table-summary