Ensure minutes have 2 digits
The previous change was stripping leading zeros from _every part_ of the
datetime used to fill in the field. If the minutes where in the single
digits, this was resulting in a time with a single minutes digit, which
was falling foul of the validation.

This commit ensures the minutes are padded with zeros again, to ensure a
double digit. Also, only the very first leading zero is removed from any
datetime part, to ensure the hours do not end up empty if the time used
is midnight.