Update SortableTable instantiation
The component now accepts the table component as the first argument,
instead of it being part of the options.

A tweak to the CSS hides the newly added SVG arrows, as these have been
created with CSS for server-side and client-side rendered sortable
tables already: this provides consistency between the designs for the 2
implementations.