Fully remove workflow manager role
This removes any uses of the workflow manager role -- mostly used to
display teh checkbox on the user management page.

This also removes tests for the correct permission on whether to show
the list of placement on a premises page: the workflow manager role was
the only role with the `cas1_premises_view` permission but without the
`cas1_space_booking_list` one, so this condition is now redundant.

Finally, a couple of bits of content still referring to the role have
been updated.