Update link to placement
Parole placement requests show a banner message, which, if the placement
request has a linked booking, shows a link to changing the booking date.
This link had not been updated to point to the new placement page. This
commit adds an integration test to cover that scenario, and fixes the
issue.