Use specific premises in E2E tests
As part of the removal of legacy bookings, we want to convert all premises in the dev/test environments to support space bookings. The end-to-end tests currently pick the first available premises on the space search screen. Previously this was ‘SWSC Men Premises 1’, but is now ‘NE Men Premises 1’, which doesn’t have any associated key workers assigned, leading to test failures further on in the tests.

This commit changes the E2E logic to always use ‘SWSC Men Premises 1’ for bookings