Remove skipped deleted user search test
This test claims that users are soft deleted and findable using the
search form on deployed env, but not locally. This does not seem to be
the case that deleted users are findable on deployed envs, so this test
does not cover existing behaviour, locally or elsewhere, and it can
therefore be removed.