Replace example service with provider service
The example service was useful as a proof of concept for the UI communicating with the API. It can now be removed.

In its place we have a provider service which will return data from the provider client relating to providers, including team data via the `getTeams()` method. This layer will be more useful in cases where we need to handle or manipulate data.