Update content on find a session page
There are a few content tweaks here, including labelling ‘Provider’ as ‘Region’ for users. This is at odds with its name in the rest of the UI and API codebase and something we may wish to change when we have slightly more confidence in the correct language.

I have also set the session table to conditionally render if there are session results passed into the template, to avoid confusion when the search page first loads.