Format times in sessions results
The API is returning local time as a string, so have added a set of utilities that:
- validates that it is indeed in the right format,
- and strips any data about seconds to display time as HH:MM.

This assumes that we want the time in the same timezone, clock type (12/24hr) as in the orginal response.

Updating the integration test allocation mock to return a time more akin to the format currently being returned.