Refactor signInStep to use page object model
As we start building out the test suite it would be useful to use page object models to aid resuability and readability of our tests.

Have introduced a basic HomePage model to use across steps. The sign in step can return this for use in the next step (as we expect the signed in user to be on the homepage).