Add data access for appointments
This adds a client and service to fetch appointment data from the Community Payback API.

While appointmentId is a number on the Dto, I am passing it around as a string
here. This is because express routes, the `static-path` library and REST API
all handle route parameters as a string, so it seems pointless to convert it to
a number in the controller, only to convert it back to a string in the client.