Add check project details view
This will be the first page of the update appointment form

The `AppointmentDto` we are using for the data on this page does not yet have
properties for all fields so leaving some rows in the summary list blank for
now.

We can add a continue/start button when we have added the next step.

We may also need to add a 'Supervising officer' input on this page, once we
have the required API endpoint to get a list of supervising officers for a
team.