Add placement selector to change placement flow
This commit introduces a new placement selector, which allows choosing
the correct placement to update for placement requests with multiple
placements. If the request only has one placement, the user is
redirected to the change placement page directly.

This commit also includes a refactor of the Space Booking factory, to
produce more realistic results with a set of dates that do not clash
with each other.