Add Offender class
The API response returns a different type of offender depending on its type:
e.g. whether it is a limited access offender with restricted details or not.

This isn't very straightforward to handle in our application. We would always
have to check the 'objectType' property and cast it to a `OffenderFullDto` to
access its information, or even to check how the offender should be handled.

We know that a limited access offender will have a blank name, and some actions
will be restricted so I have opted to create a generic interface for this in
our UI with a new model.

The constructor will handle setting the full name based on the orginal object
type. A boolean `isLimited` property on the instantiated class can be uses to
show or hide functionality.