Ensure selected keyworker has assigned placements
Because the keyworker factory assigns a random number of current and
upcoming bookings between 0 and 20, there was a 1 in 21 chance that
keyworker selected by the test to filter on had no upcoming bookings
assigned, and would therefore not show in the dropdown, resulting in a
test failure.

This ensures the first keyworker in the fixture has a count of upcoming
and current bookings more than 0.