Fix and improve summary list matcher
The previous iteration would assert that any value existed against _all_
`<dd>` elements. This was incredibly inefficient on pages with big
summary lists, and could potentially return a match for a definition
not related to the summary list item being checked.

This version now ensures the definition being checked is the actual
sibling of the key that has been found. It also ensures that, should a
value be empty (or only whitespace), then the definition element does
match this.